+++ server/configs/cas.js
@@ -3,7 +3,7 @@ var CASAuthentication = require('node-cas-authentication')
 var cas = new CASAuthentication({
   cas_url: process.env.CAS_URL,
   service_url: process.env.CAS_SERVICE_URL,
-  cas_version: '3.0',
+  cas_version: '2.0',
   renew: false,
   is_dev_mode: process.env.CAS_DEV_MODE === 'true',
   dev_mode_user: process.env.CAS_DEV_USER,
