+++ server/routes/auth.js
@@ -44,6 +44,7 @@ router.use(requestLogger)
  *         description: user is logged in, redirect to homepage
  */
 router.get('/login', refreshToken, async function (req, res, next) {
+  console.log("session:", req.session)
   if (!req.session.user_id) {
     let email = ''
     if (req.query.email && process.env.FORCE_AUTH === 'true') {
